+++ lapce-ui/src/palette.rs
@@ -2,16 +2,17 @@ use std::{path::Path, sync::Arc};
 
 use druid::{
     kurbo::Rect,
-    piet::{Svg, Text, TextAttribute, TextLayoutBuilder},
-    BoxConstraints, Color, Command, Data, Env, Event, EventCtx, FontWeight,
+    piet::{PietTextLayout, Svg, Text, TextAttribute, TextLayoutBuilder},
+    theme, BoxConstraints, Color, Command, Data, Env, Event, EventCtx, FontWeight,
     LayoutCtx, LifeCycle, LifeCycleCtx, Modifiers, PaintCtx, Point, RenderContext,
     Size, Target, UpdateCtx, Widget, WidgetExt, WidgetId, WidgetPod,
 };
+use lapce_core::mode::{Mode, Modes};
 use lapce_data::{
     command::{LapceUICommand, LAPCE_COMMAND, LAPCE_UI_COMMAND},
     config::{LapceConfig, LapceTheme},
     data::{LapceTabData, LapceWorkspaceType},
-    keypress::KeyPressFocus,
+    keypress::{KeyPress, KeyPressFocus},
     list::ListData,
     palette::{
         PaletteItem, PaletteItemContent, PaletteListData, PaletteStatus,
@@ -272,8 +273,16 @@ impl Widget<LapceTabData> for PaletteContainer {
     ) {
         self.input.event(ctx, event, data, env);
 
+        let mode = data.mode();
         let palette = Arc::make_mut(&mut data.palette);
         palette.list_data.update_data(data.config.clone());
+        // Update the stored data on `PaletteListData`, this is so that they can be used
+        // during painting of the list
+        palette.list_data.data.workspace = Some(data.workspace.clone());
+        palette.list_data.data.keymaps =
+            Some(data.keypress.commands_with_keymap.clone());
+        palette.list_data.data.mode = Some(mode);
+
         self.content.event(ctx, event, &mut palette.list_data, env);
 
         self.preview.event(ctx, event, data, env);
@@ -504,6 +513,7 @@ struct PaletteItemPaintInfo {
     text_indices: Vec<usize>,
     hint: String,
     hint_indices: Vec<usize>,
+    key: Option<(KeyPress, Option<KeyPress>)>,
 }
 
 impl PaletteItemPaintInfo {
@@ -516,6 +526,7 @@ impl PaletteItemPaintInfo {
             text_indices,
             hint: String::new(),
             hint_indices: Vec::new(),
+            key: None,
         }
     }
 }
@@ -525,7 +536,7 @@ impl ListPaint<PaletteListData> for PaletteItem {
         &self,
         ctx: &mut PaintCtx,
         data: &ListData<Self, PaletteListData>,
-        _env: &Env,
+        env: &Env,
         line: usize,
     ) {
         let PaletteItemPaintInfo {
@@ -535,6 +546,7 @@ impl ListPaint<PaletteListData> for PaletteItem {
             text_indices,
             hint,
             hint_indices,
+            key,
         } = match &self.content {
             PaletteItemContent::File(path, _) => {
                 file_paint_items(path, &self.indices, data)
@@ -582,6 +594,7 @@ impl ListPaint<PaletteListData> for PaletteItem {
                     text_indices,
                     hint,
                     hint_indices,
+                    key: None,
                 }
             }
             PaletteItemContent::WorkspaceSymbol {
@@ -624,8 +637,37 @@ impl ListPaint<PaletteListData> for PaletteItem {
                     .kind
                     .desc()
                     .map(|m| m.to_string())
-                    .unwrap_or_else(|| "".to_string());
-                PaletteItemPaintInfo::new_text(text, self.indices.to_vec())
+                    .unwrap_or_else(String::new);
+
+                let key = if let Some(keymaps) = data.data.keymaps.as_ref() {
+                    // The internal name of the command
+                    let command_str = command.kind.str();
+                    let mode = Modes::from(data.data.mode.unwrap_or(Mode::Normal));
+                    // Get the keymaps that match this command (if there are any) and join
+                    // at most two of the keypresses in them for the hint text.
+                    let mut keys = keymaps
+                        .iter()
+                        .filter(|keymap| keymap.command == command_str)
+                        .filter(|keymap| {
+                            keymap.modes.is_empty() || keymap.modes.contains(mode)
+                        })
+                        .flat_map(|keymap| &keymap.key)
+                        .take(2)
+                        .cloned();
+                    keys.next().map(|key| (key, keys.next()))
+                } else {
+                    Default::default()
+                };
+
+                PaletteItemPaintInfo {
+                    svg: None,
+                    svg_color: None,
+                    text,
+                    text_indices: self.indices.to_vec(),
+                    hint: String::new(),
+                    hint_indices: Vec::new(),
+                    key,
+                }
             }
             PaletteItemContent::ColorTheme(theme) => PaletteItemPaintInfo::new_text(
                 theme.to_string(),
@@ -756,6 +798,62 @@ impl ListPaint<PaletteListData> for PaletteItem {
         let y = line_height * line as f64 + text_layout.y_offset(line_height);
         let point = Point::new(x, y);
         ctx.draw_text(&text_layout, point);
+
+        // TODO: make sure that the main text and the keymaps don't overlap
+        // This isn't currently an issue for any added commands, but could become one
+
+        if let Some((key0, key1)) = key {
+            let (key0_right, key0_items) =
+                key0.paint(ctx, Point::ZERO, &data.config);
+            let (rightmost, key1_items) = if let Some(key1) = key1 {
+                let (key1_right, items) = key1.paint(
+                    ctx,
+                    Point::new(key0_right.x + 5.0, 0.0),
+                    &data.config,
+                );
+                (key1_right.x, Some(items))
+            } else {
+                (key0_right.x, None)
+            };
+
+            let width = ctx.size().width;
+            let x = width
+                - rightmost
+                - env.get(theme::SCROLLBAR_WIDTH)
+                - env.get(theme::SCROLLBAR_PAD);
+
+            let origin = Point::new(x, y);
+            paint_key_parts(ctx, &data.config, origin, key0_items);
+
+            if let Some(key1_items) = key1_items {
+                paint_key_parts(ctx, &data.config, origin, key1_items);
+            }
+        }
+    }
+}
+
+fn paint_key_parts(
+    ctx: &mut PaintCtx,
+    config: &LapceConfig,
+    origin: Point,
+    items: Vec<(Option<Rect>, PietTextLayout, Point)>,
+) {
+    for (rect, text_layout, text_layout_pos) in items {
+        if let Some(mut rect) = rect {
+            rect.x0 += origin.x;
+            rect.x1 += origin.x;
+            rect.y0 += origin.y;
+            rect.y1 += origin.y;
+            ctx.stroke(
+                rect,
+                config.get_color_unchecked(LapceTheme::LAPCE_BORDER),
+                1.0,
+            );
+        }
+        ctx.draw_text(
+            &text_layout,
+            Point::new(text_layout_pos.x + origin.x, text_layout_pos.y + origin.y),
+        );
     }
 }
 
@@ -809,6 +907,7 @@ fn file_paint_symbols(
         text_indices,
         hint,
         hint_indices,
+        key: Default::default(),
     }
 }
 
@@ -862,5 +961,6 @@ fn file_paint_items(
         text_indices,
         hint: folder,
         hint_indices,
+        key: Default::default(),
     }
 }