+++ lapce-ui/src/status.rs
@@ -8,8 +8,8 @@ use lapce_core::mode::Mode;
 use lapce_data::{
     command::{CommandKind, LapceCommand, LapceWorkbenchCommand, LAPCE_COMMAND},
     config::{LapceConfig, LapceIcons, LapceTheme},
-    data::{FocusArea, LapceTabData},
-    panel::{PanelContainerPosition, PanelKind},
+    data::LapceTabData,
+    panel::PanelContainerPosition,
 };
 
 use crate::tab::LapceIcon;
@@ -330,16 +330,7 @@ impl Widget<LapceTabData> for LapceStatus {
         let mut _right = 0.0;
 
         if data.config.core.modal {
-            let mode = if data.focus_area == FocusArea::Panel(PanelKind::Terminal) {
-                data.terminal
-                    .terminals
-                    .get(&data.terminal.active_term_id)
-                    .map(|terminal| terminal.mode)
-            } else {
-                data.main_split.active_editor().map(|e| e.cursor.get_mode())
-            };
-
-            let (mode, color) = match mode.unwrap_or(Mode::Normal) {
+            let (mode, color) = match data.mode() {
                 Mode::Normal => ("Normal", LapceTheme::STATUS_MODAL_NORMAL),
                 Mode::Insert => ("Insert", LapceTheme::STATUS_MODAL_INSERT),
                 Mode::Visual => ("Visual", LapceTheme::STATUS_MODAL_VISUAL),