+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features/Changes
 - [#1720](https://github.com/lapce/lapce/pull/1720): Display signature/parameter information from LSP
+- [#1723](https://github.com/lapce/lapce/pull/1723): In the palette, display the keybind for a command adjacent to it
 
 ### Bug Fixes